♻️ refactor: app 디렉토리리 비우기
